Big changes
Still in the baby steps. Getting each piece to work individually is less and less of a priority. Figuring out how to get these individual pieces to talk to each other. Some CAD work for the chassis has been done, but that's slow going because of my lack of CAD knowledge. Next few steps will be how to take pictures for later stages and to finish the cad then cut the chassis.