Sensors built and working
So you'll see the Bumble2.py file in currently running code because that is what is running on the robot right now. It does detect stairs reliably, right now the focus is back on hardware for the power supply because this little robot sucks through a lot of power. However the prototype battery management system is underway. So its getting better. I want to have it rolling around the house within 7 days and the battery management within 14. So that's where the project stands right now.