Breaking out code to a few files
This is just a step forward for the code and making it cleaner by formalizing the sensors and motors into classes. More will be cleaned up in the future and some of the kinks ironed out. Right now there is an issue where the sensors take too long to sense that there is no stair so I need to do some testing and see whats up with that. probably some timing data or something.